+++ src/injectable.ts
@@ -1,6 +1,6 @@
 /**
  * Indicates that the decorated class can have it's constructor parameters injected.
  */
-export function injectable(constructor: Function): void {   // tslint:disable-line:ban-types
+export function injectable(constructor: Function): void {
     // no-op
 }