+++ src/types.ts
@@ -1,11 +1,6 @@
 
-// tslint:disable-next-line:interface-name
 export interface Constructor<T> {
     new(...args: any[]): T;
-}
-
-export interface IDictionary<T> {
-    [key: string]: T;
 }
 
 export type Factory<T> = () => T;