+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Changed
 
+- **BREAKING**: `ResolveOptions.params` is now a `Map` instead of a plain object.
 - Change dev stack to Babel, Jest and ESLint.
 
 ## [1.1.0 - 2018-06-29](https://github.com/alonrbar/peppermint-di/tree/v1.1.0)